5.1.22 Перенесет логику в контроллер и подключит его в main
